requirements.txt
astroid==2.4.2
autopep8==1.5.5
Django==2.2
django-ckeditor==6.0.0
django-decouple==2.1
django-etc==1.3.0
django-hitcount==1.3.3
django-js-asset==1.2.2
djangorestframework==3.12.2
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
pycodestyle==2.6.0
pylint==2.6.0
python-slugify==4.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
whitenoise==5.2.0
wrapt==1.12.1
